Moving update company profile to admin api. This operation requires user + App credentials
