+++ src/main/java/org/springframework/social/partnercenter/api/customer/AdminCustomerOperations.java
@@ -5,5 +5,6 @@ import org.springframework.social.partnercenter.api.PartnerCenterResponse;
 
 public interface AdminCustomerOperations extends CustomerOperations{
 	ResponseEntity<CustomerBillingProfile> getBillingProfile(String customerId);
+	ResponseEntity<CustomerBillingProfile> updateBillingProfile(String customerId, String etag, CustomerBillingProfile billingProfile);
 	ResponseEntity<PartnerCenterResponse<Customer>> getCompanyByDomain(int size, String filter);
 }
