+++ src/main/java/org/springframework/social/partnercenter/api/customer/CustomerOperations.java
@@ -15,6 +15,5 @@ public interface CustomerOperations extends PagingResourceOperations<Customer>{
 	ResponseEntity<PartnerCenterResponse<Customer>> getCompanyByCompanyName(int size, String filter);
 	ResponseEntity<CustomerListResponse> getList(int size);
 	ResponseEntity<CustomerCompanyProfile> getCustomersCompanyProfile(String customerId);
-    ResponseEntity<CustomerBillingProfile> updateBillingProfile(String customerId, String etag, CustomerBillingProfile billingProfile);
     ResponseEntity<Subscription> updateFriendlyName(String customerTenantId, String subscriptionId, String nickname);
 }
